chore: export promisify types
- required for using types for the LndSerivce outside the plugin itself
- move all types to `types/` folder
